Create collection should use the same zk client when checking cluster state.
Previously, the collection could be seen as existing after the initial zk
collection creation using the solr cloud manager zk reader. But after creating the
replicas, it could still be missing in the zk reader from the OverseerCollectionMessageHandler
due to watches triggering and being handled at different times resulting in different local
state. Make sure we use the same cluster state reader throughout for a consistent view
of the data.