Cache Plugin Loading: Use resolved plugin classpath (#65)
Plugin class paths take the form of name space, like "xyz" and a class path "foo.bar.baz" which forms "xyz:foo.bar.baz". Plugin loading takes the previous string and breaks it apart into a package and classpath component. This updates cache instantiation to use the class path portion of the configured cache if it is present